requirements.txt
python-slugify==4.0.0
requests==2.23.0
